requirements.txt
anyio==4.2.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.2
idna==3.6
inflection==0.5.1
packaging==23.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.0
